Update JWT embed properties
This commit updates the embed to not use the On Demand Access properties
as part of the JWT since they require a license upgrade. The idea here
is that Tableau visualizations would be embedded as appropriate in the
web application, and they would otherwise only be available to selected
users in Tableau.

I had to change the URL for the visualization, perhaps that changed over
time? We will obviously be updating these to reference the actual
visualizations we care about.

Also, it includes some dependency updates.